requirements.txt
asyncio
aiosqlite
sqlalchemy
fastapi
strawberry-graphql[fastapi]
uvicorn